Add files via upload
The final version of the HMI team's work intented for use for the TU Summer Practice. Changes from previous version include: a fix in the calculations and transformation of units, the removal of all debug features, leaving only the main code. Upon loading the window it should connect to the server(VIP) via a webSocket and start recieving and visualizing all the necessary data.
13.07.2019------03:10